Store accessible interfaces in the model
Ensures that, whenever an accessible interface is requested, the same
iterface is returned each time rather than a new one being created.
This may result in more accurate speech in some situtations, such as
when an item changes state (e.g. from selected to not selected).

So far, the only noticable benefit occurs on Windows, where Narrator is
now able to report item numbers correctly (e.g. "1 of 28", "2 of 28",
etc.). Previously it would always say "0 of 28" regardless of which
item was selected.